-tested and fixed new camouflage types and ammo weights -fixed localization to read in XML file
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@400 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Soldier Create.cpp
@@ -457,10 +457,10 @@ SOLDIERTYPE* TacticalCreateSoldier( SOLDIERCREATE_STRUCT *pCreateStruct, UINT8 *
 					if ( i != NO_SLOT && Random( 5 ) < SoldierDifficultyLevel( &Soldier ))
 					{
 						// start camouflaged
-						Soldier.bCamo = Item[i].camobonus;
-						Soldier.urbanCamo = Item[i].urbanCamobonus;
-						Soldier.desertCamo = Item[i].desertCamobonus;
-						Soldier.snowCamo = Item[i].snowCamobonus;
+						Soldier.bCamo = (INT8)Item[i].camobonus;
+						Soldier.urbanCamo = (INT8)Item[i].urbanCamobonus;
+						Soldier.desertCamo = (INT8)Item[i].desertCamobonus;
+						Soldier.snowCamo = (INT8)Item[i].snowCamobonus;
 					}
 				}
 			}
